requirements.txt
streamlit==1.31.0
requests==2.31.0
python-dateutil==2.8.2
PyYAML==6.0.1
